+++ logo/logo.py
@@ -60,10 +60,10 @@ class Thumbnail(GraphScene):
             triangle.scale(0.1)
         
         # 
-        x_label_p1 = MTex("a")
-        output_label_p1 = MTex("f(a)")
-        x_label_p2 = MTex("b")
-        output_label_p2 = MTex("f(b)")
+        x_label_p1 = Tex("a")
+        output_label_p1 = Tex("f(a)")
+        x_label_p2 = Tex("b")
+        output_label_p2 = Tex("f(b)")
         v_line_p1 = get_v_line(input_tracker_p1)
         v_line_p2 = get_v_line(input_tracker_p2)
         h_line_p1 = get_h_line(input_tracker_p1)
@@ -170,7 +170,7 @@ class Thumbnail(GraphScene):
         # adding manim
         picture = Group(*self.mobjects)
         picture.scale(0.6).to_edge(LEFT, buff=SMALL_BUFF)
-        manim = MTexText("Manim").set_height(1.5) \
+        manim = TexText("Manim").set_height(1.5) \
                                     .next_to(picture, RIGHT) \
                                     .shift(DOWN * 0.7)
         self.add(manim)
