+++ manimlib/mobject/svg/drawings.py
@@ -44,8 +44,8 @@ from manimlib.mobject.geometry import Square
 from manimlib.mobject.mobject import Mobject
 from manimlib.mobject.numbers import Integer
 from manimlib.mobject.svg.svg_mobject import SVGMobject
-from manimlib.mobject.svg.mtex_mobject import MTex
-from manimlib.mobject.svg.mtex_mobject import MTexText
+from manimlib.mobject.svg.tex_mobject import Tex
+from manimlib.mobject.svg.tex_mobject import TexText
 from manimlib.mobject.svg.special_tex import TexTextFromPresetString
 from manimlib.mobject.three_dimensions import Prismify
 from manimlib.mobject.three_dimensions import VCube
@@ -427,7 +427,7 @@ class Bubble(SVGMobject):
         return self.content
 
     def write(self, *text):
-        self.add_content(MTexText(*text))
+        self.add_content(TexText(*text))
         return self
 
     def resize_to_content(self, buff=0.75):
